refactor(ai-agent): merge acp_service into acp_agent_service
Fold CLI detection helpers (detect_cli, health_check, get_env,
test_custom_agent) into acp_agent_service alongside the existing
per-session state facade, and drop the now-empty acp_service module.
