+++ crates/aionui-ai-agent/src/lib.rs
@@ -5,7 +5,6 @@ pub mod acp_error;
 pub mod acp_protocol;
 pub mod acp_routes;
 pub mod acp_runtime_snapshot;
-pub mod acp_service;
 pub mod agent_manager;
 pub mod agent_registry;
 pub mod agent_routes;
